# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "a http server in front of mmcli",
        "title": "mmcli server",
        "contact": {},
        "license": {
            "name": "bsd"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8743",
    "paths": {
        "/location": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "location"
                ],
                "summary": "Get location",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Modem-Id",
                        "name": "modem",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/modem/": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modem"
                ],
                "summary": "List modems",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/modem/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modem"
                ],
                "summary": "Modem info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Modem-Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "utils"
                ],
                "summary": "ping",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sms": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms"
                ],
                "summary": "Send SMS",
                "parameters": [
                    {
                        "description": "Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SmsRequestData"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Modem-Id",
                        "name": "modem",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/sms/": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms"
                ],
                "summary": "List SMS messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Modem-Id",
                        "name": "modem",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/sms/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms"
                ],
                "summary": "List SMS messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SMS-Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Modem-Id",
                        "name": "modem",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms"
                ],
                "summary": "Delete a single SMS",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SMS-Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "api.SmsRequestData": {
            "type": "object",
            "required": [
                "number",
                "text"
            ],
            "properties": {
                "number": {
                    "type": "string",
                    "example": "+431234567890"
                },
                "text": {
                    "type": "string",
                    "maxLength": 160,
                    "example": "Ping"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}